requirements.txt
opencv-python
pytorch-lightning
hydra-core
numpy
scikit-learn
matplotlib
transformers
wandb